Init ADC AFS

+++ AFS/inc/ADC.h
@@ -1,2 +1,3 @@
-int Get_Result (void);
+#include "stm8s.h"
+uint16_t Get_Result (void);
 void ADC_Config (void);
